move tooltip positioning to a hook
